regex added and fix menu
