--- src/main/java/com/infoshareacademy/Entity/User.java
@@ -1,10 +1,0 @@
-package com.infoshareacademy.Entity;
-
-public abstract class User {
-    protected String userLogin;
-    protected String userPassword;
-    protected String name;
-    protected String phoneNumber;
-    protected Long id;
-    protected byte userType;
-}
