--- src/main/java/com/infoshareacademy/Utilities/RoomSearch.java
@@ -1,4 +1,0 @@
-package com.infoshareacademy.Utilities;
-
-public class RoomSearch {
-}
