--- src/main/java/com/infoshareacademy/View/Screen.java
@@ -1,11 +1,0 @@
-package com.infoshareacademy.View;
-
-public abstract class Screen {
-
-    public static void clearScreen() {
-        System.out.println();
-        System.out.println();
-        System.out.println();
-        System.out.println();
-    }
-}
