+++ src/main/java/com/infoshareacademy/file/operation/FilesNames.java
@@ -1,0 +1,11 @@
+package com.infoshareacademy.file.operation;
+
+public class FilesNames {
+
+    public static final String TENANTS_JSON = "src/main/resources/tenants.json";
+    public static final String OWNERS_JSON = "src/main/resources/owners.json";
+    public static final String ROOMS_JSON = "src/main/resources/rooms.json";
+
+
+
+}
