+++ src/main/java/com/infoshareacademy/utilities/RoomEdit.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.Utilities;
+package com.infoshareacademy.utilities;
 
 public class RoomEdit {
