+++ src/main/java/com/infoshareacademy/utilities/RoomSearch.java
@@ -1,0 +1,4 @@
+package com.infoshareacademy.utilities;
+
+public class RoomSearch {
+}
