+++ src/main/resources/owners.json
@@ -20,6 +20,34 @@
       "surname": "madrak",
       "email": "madr",
       "id": "461da64b-3b4a-4e7e-b5eb-96be8ebd244a"
+    },
+    {
+      "login": "Aneczka",
+      "name": "Ania",
+      "surname": "Ania",
+      "email": "Ania@gmail.com",
+      "id": "d2e1fa52-2c6b-4483-9768-a93f526f1dc2"
+    },
+    {
+      "login": "AniaB",
+      "name": "546",
+      "surname": "546",
+      "email": "Ania@545.pl",
+      "id": "6db050b7-2dd7-43e5-a25f-e707c618c472"
+    },
+    {
+      "login": "IwonaB",
+      "name": "Basia",
+      "surname": "Kowalska",
+      "email": "Iwona.kowalska@onet.pl",
+      "id": "5ffddb42-09ff-485c-bdf3-18dbab706bad"
+    },
+    {
+      "login": "Ania",
+      "name": "Ania",
+      "surname": "ANia",
+      "email": "Ania@Ania.pl",
+      "id": "59ceff95-35f1-414d-9297-d52550266f0f"
     }
   ]
 }
