+++ src/main/resources/tenants.json
@@ -38,6 +38,13 @@
       "email": "gol@dlapolski",
       "id": "f315ad8a-b87f-4b81-90da-b1e60f60642d",
       "AssigmentId": "5dec05e0-ce27-42c4-b168-c84d4c16edc1"
+    },
+    {
+      "login": "Ania",
+      "name": "Ania",
+      "surname": "Ania",
+      "email": "Aneczka@aneczka.com",
+      "id": "571a7e9b-8e3d-4442-87a1-bc6446fd1fbc"
     }
   ]
 }
